refactor: overhaul authentication components and streamline user flow
- Removed AuthForm and Accounts components to simplify the authentication process.
- Integrated login and signup functionalities directly into LoginPage and SignupPage components.
- Enhanced user experience with loading indicators during authentication actions.
- Updated useAuthStore to manage authentication state and actions more effectively.
- Improved error handling and user feedback through toast notifications.
