+++ frontend/src/store/useAuthStore.js
@@ -1,5 +1,6 @@
 import { axiosInstance } from "../lib/axios";
 import { create } from "zustand";
+import { showToast } from "../util/toast";
 
 export const useAuthStore = create((set) => ({
   authUser: null,
@@ -18,6 +19,68 @@ export const useAuthStore = create((set) => ({
       set({ authUser: null });
     } finally {
       set({ isCheckingAuth: false });
+    }
+  },
+
+  signUp: async (data, navigate) => {
+    set({ isSigningUp: true });
+    try {
+      const res = await axiosInstance.post("/auth/sign-up", data);
+      set({ authUser: res.data });
+      showToast({
+        description: "Account Created Successfully!",
+        status: "success",
+      });
+      navigate("/");
+    } catch (error) {
+      console.log("Error Message: ", error);
+      showToast({
+        description: error.response.data.message,
+        status: "error",
+      });
+    } finally {
+      set({ isSigningUp: false });
+    }
+  },
+
+  login: async (data, navigate) => {
+    set({ isLoggingIn: true });
+    try {
+      const res = await axiosInstance.post("/auth/login", data);
+      set({ authUser: res.data });
+      showToast({
+        description: "Logged in Succesfully!",
+        status: "success",
+      });
+      navigate("/");
+    } catch (error) {
+      console.log("Error Message: ", error);
+      showToast({
+        description: error.response.data.message,
+        status: "error",
+      });
+    } finally {
+      set({ isLoggingIn: false });
+    }
+  },
+
+  logout: async (navigate) => {
+    set({ isLoggingOut: true });
+    try {
+      const res = await axiosInstance.post("/auth/logout");
+      navigate("/login");
+      showToast({
+        description: "Logged out Succesfully!",
+        status: "success",
+      });
+    } catch (error) {
+      console.log("Error Message: ", error);
+      showToast({
+        description: error.response.data.message,
+        status: "error",
+      });
+    } finally {
+      set({ isLoggingOut: false });
     }
   },
 }));
